Partial cleanup of test_smirnoff. Moved some tests and removed redundant ones.
