+++ openforcefield/tests/test_smirnoff.py
@@ -243,295 +243,6 @@ ffxml_MDL_contents = u"""\
 # Test various components
 #
 
-class TestForceField:
-    """Test capabilities of ForceField and its Python API.
-    """
-
-    # TODO: Can we refine `forcefield.get_handler()` and `hander.add_parameter()` interface?
-
-    def create_forcefield_via_api(self):
-        """Create a minimal ForceField via the API and test object data model is correctly populated.
-        """
-        # Create a ForceField
-        forcefield = ForceField(version='1.0', aromaticity_model="MDL")
-        # Add bonds
-        # TODO: I don't like this API for adding parameter blocks; can we do better?
-        handler = forcefield.get_handler('Bonds', potential="harmonic")
-        length_unit = unit.angstroms
-        k_unit = unit.kilocalories_per_mole / unit.angstrom
-        # TODO: Should the API automagically handle the case where parameters are provided as strings?
-        handler.add_parameter(smirks="[#6X4:1]-[#6X4:2]", length=1.526 * length_unit, k=620.0 * k_unit)
-        handler.add_parameter(smirks="[#6X4:1]-[#1:2]", length=1.090 * length_unit, k=680.0 * k_unit)
-        assert_equal(forcefield.forces['Bonds'].parameters[0].length, 1.526 * length_unit)
-        assert_equal(forcefield.forces['Bonds'].parameters[1].k, 680.0 * length_unit)
-        assert_equal(forcefield.forces['Bonds'].parameters["[#6X4:1]-[#1:2]"].length, 1.090 * length_unit)
-        # Add angles
-        handler = forcefield.get_handler('Angles', potential="harmonic")
-        angle_unit = unit.degrees
-        k_unit = unit.kilocalories_per_mole / unit.radian**2
-        handler.add_parameter(smirks="[a,A:1]-[#6X4:2]-[a,A:3]", angle=109.50 * angle_unit, k=100.0 * k_unit)
-        handler.add_parameter(smirks="[#1:1]-[#6X4:2]-[#1:3]", angle=109.50 * angle_unit, k=70.0 * k_unit)
-        assert_equal(forcefield.forces['Angles'].parameters[0].angle, 109.50 * angle_unit)
-        assert_equal(forcefield.forces['Angles'].parameters[0].k, 100.0 * k_unit)
-        # Add proper torsions
-        handler = forcefield.get_handler('ProperTorsions', potential="charmm")
-        phase_unit = unit.degrees
-        k_unit = unit.kilocalories_per_mole
-        handler.add_parameter(
-            smirks="[a,A:1]-[#6X4:2]-[#6X4:3]-[a,A:4]",
-            idivf1=9,
-            periodicity1=3,
-            phase1=0.0 * phase_unit,
-            k1=1.40 * k_unit)
-        handler.add_parameter(
-            smirks="[#6X4:1]-[#6X4:2]-[#8X2:3]-[#6X4:4]",
-            idivf1=1,
-            periodicity1=3,
-            phase1=0.0 * phase_unit,
-            k1=0.383 * k_unit,
-            idivf2=1,
-            periodicity2=2,
-            phase2=180.0 * phase_unit,
-            k2=0.1 * k_unit)
-        assert hasattr(forcefield.forces['ProperTorsions'].parameters[0], 'k1')
-        assert not hasattr(forcefield.forces['ProperTorsions'].parameters[0], 'k2')
-        assert hasattr(forcefield.forces['ProperTorsions'].parameters[1], 'k1')
-        assert hasattr(forcefield.forces['ProperTorsions'].parameters[2], 'k2')
-        assert_equal(forcefield.forces['ProperTorsions'].parameters["[#6X4:1]-[#6X4:2]-[#8X2:3]-[#6X4:4]"].idivf1, 1)
-        assert_equal(
-            forcefield.forces['ProperTorsions'].parameters["[#6X4:1]-[#6X4:2]-[#8X2:3]-[#6X4:4]"].periodicity1, 3)
-        assert_equal(forcefield.forces['ProperTorsions'].parameters["[#6X4:1]-[#6X4:2]-[#8X2:3]-[#6X4:4]"].phase1,
-                     0.0 * phase_unit)
-        assert_equal(forcefield.forces['ProperTorsions'].parameters["[#6X4:1]-[#6X4:2]-[#8X2:3]-[#6X4:4]"].k1,
-                     0.383 * k_unit)
-        assert_equal(forcefield.forces['ProperTorsions'].parameters["[#6X4:1]-[#6X4:2]-[#8X2:3]-[#6X4:4]"].idivf2, 1)
-        assert_equal(
-            forcefield.forces['ProperTorsions'].parameters["[#6X4:1]-[#6X4:2]-[#8X2:3]-[#6X4:4]"].periodicity2, 2)
-        assert_equal(forcefield.forces['ProperTorsions'].parameters["[#6X4:1]-[#6X4:2]-[#8X2:3]-[#6X4:4]"].phase2,
-                     180.0 * phase_unit)
-        assert_equal(forcefield.forces['ProperTorsions'].parameters["[#6X4:1]-[#6X4:2]-[#8X2:3]-[#6X4:4]"].k3,
-                     0.1 * k_unit)
-        # Add improper torsions
-        handler = forcefield.get_handler('ImproperTorsions', potential="charmm", default_idivf="auto")
-        phase_unit = unit.degrees
-        k_unit = unit.kilocalories_per_mole
-        handler.add_parameter(
-            smirks="[*:1]~[#6X3:2](=[#7X2,#7X3+1:3])~[#7:4]",
-            k1=10.5 * k_unit,
-            periodicity1=2,
-            phase1=180.0 * phase_unit)
-        assert_equal(forcefield.forces['ImproperTorsions'].parameters[0].k1, 10.5 * k_unit)
-        assert_equal(forcefield.forces['ImproperTorsions'].parameters[0].periodicity1, 2)
-        assert_equal(forcefield.forces['ImproperTorsions'].parameters[0].phase1, 180.0 * phase_unit)
-        # Add vdW
-        handler = forcefield.get_handler(
-            'vdW',
-            potential="Lennard-Jones-12-6",
-            combining_rules="Loentz-Berthelot",
-            switch=8.0 * unit.angstroms,
-            cutoff=9.0 * unit.angstroms,
-            long_range_dispersion="isotropic")
-        sigma_unit = unit.angstroms
-        epsilon_unit = unit.kilocalories_per_mole
-        handler.add_parameter(smirks="[#1:1]", sigma=1.4870 * sigma_unit, epsilon=0.0157 * epsilon_unit)
-        handler.add_parameter(smirks="[#1:1]-[#6]", sigma=1.4870 * sigma_unit, epsilon=0.0157 * epsilon_unit)
-        assert_equal(forcefield.forces['vdW'].parameters[0].sigma, 1.4870 * sigma_unit)
-        assert_equal(forcefield.forces['vdW'].parameters[0].epsilon, 0.0157 * epsilon_unit)
-        # Add ChargeIncrementModel
-        handler = forcefield.get_handler(
-            'ChargeIncrementModel',
-            number_of_conformers=10,
-            quantum_chemical_method="AM1",
-            partial_charge_method="CM2")
-        charge_unit = unit.elementary_charge
-        handler.add_parameter(
-            smirks="[#6X4:1]-[#6X3a:2]",
-            charge1increment=-0.0073 * charge_unit,
-            charge2increment=+0.0073 * charge_unit)
-        handler.add_parameter(
-            smirks="[N:1](H:2)(H:3)",
-            charge1increment=+0.02 * charge_unit,
-            charge2increment=-0.01 * charge_unit,
-            charge3increment=-0.01 * charge_unit)
-        assert_equal(forcefield.forces['ChargeIncrementModel'].parameters[0].smirks, "[#6X4:1]-[#6X3a:2]")
-        assert_equal(forcefield.forces['ChargeIncrementModel'].parameters["[N:1](H:2)(H:3)"].charge1increment,
-                     +0.02 * charge_unit)
-        # Add LibraryCharges
-        handler = forcefield.get_handler('LibraryCharges')
-        charge_unit = unit.elementary_charge
-        handler.add_parameter(
-            name="TIP3P",
-            smirks="[#1:1]-[#8X2H2+0:2]-[#1:3]",
-            charge1=+0.417 * charge_unit,
-            charge2=-0.834 * charge_unit,
-            charge3=+0.417 * charge_unit)
-        assert_equal(forcefield.forces['LibraryCharges'].parameters[0].name, 'TIP3P')
-        assert_equal(forcefield.forces['LibraryCharges'].parameters[0].charge1, +0.417 * charge_unit)
-        assert_equal(forcefield.forces['LibraryCharges'].parameters[0].charge2, -0.834 * charge_unit)
-        assert_equal(forcefield.forces['LibraryCharges'].parameters[0].charge3, +0.417 * charge_unit)
-        assert not hasattr(forcefield.forces['LibraryCharges'].parameters[0], 'charge4')
-
-        return forcefield
-
-    def test_create(self):
-        """Test creating a ForceFied via the Python API.
-        """
-        forcefield = self.create_forcefield_via_api()
-
-    def test_modify(self):
-        """Test modifying a ForceField via the Python API.
-        """
-        forcefield = self.create_forcefield_via_api()
-        # Modify some parameters
-        forcefield.forces['Bonds'].parameters[0].k *= 1.1
-        forcefield.forces['Bonds'].parameters[1].length += 0.01 * unit.nanometers
-        forcefield.forces['Bonds'].parameters["#6X4:1]-[#1:2]"].smirks += "~[#7]"
-        assert "#6X4:1]-[#1:2]" not in forcefield.forces['Bonds'].parameters
-        assert "#6X4:1]-[#1:2]~[#7]" in forcefield.forces['Bonds'].parameters
-
-    def test_delete_parameters(self):
-        forcefield = self.create_forcefield_via_api()
-        # Delete some parameters
-        forcefield.forces['Bonds'].parameters[0].k *= 1.1
-        forcefield.forces['Bonds'].parameters[1].length += 0.01 * unit.nanometers
-        forcefield.forces['Bonds'].parameters["#6X4:1]-[#1:2]"].smirks += "~[#7]"
-        assert "#6X4:1]-[#1:2]" not in forcefield.forces['Bonds'].parameters
-        assert "#6X4:1]-[#1:2]~[#7]" in forcefield.forces['Bonds'].parameters
-
-
-class TestXMLForceField:
-    """Test capabilities of the XML reader/writer for ForceField.
-    """
-
-    def test_attached_units():
-        from openforcefield.typing.engines.smirnoff.io import _extract_attached_units, _attach_units
-        # Test extraction of units
-        force_attrib = OrderedDict([
-            ('potential', "Lennard-Jones-12-6"),
-            ('combining_rules', "Loentz-Berthelot"),
-            ('scale12', "0.0"),
-            ('scale13', "0.0"),
-            ('scale14', "0.5"),
-            ('sigma_unit', "angstroms"),
-            ('epsilon_unit', "kilocalories_per_mole"),
-            ('switch', "8.0*angstroms"),
-            ('cutoff', "9.0*angstroms"),
-            ('long_range_dispersion', "isotropic"),
-        ])
-        stripped_attrib, attached_units = _extract_attached_units(force_attrib)
-        assert_equal(
-            list(stripped_attrib.keys()), [
-                'potential', 'combining_rules', 'scale12', 'scale13', 'scale14', 'switch', 'cutoff',
-                'long_range_dispersion'
-            ])
-        assert_equal(
-            list(stripped_attrib.values()), [
-                "Lennard-Jones-12-6", "Loentz-Berthelot", "0.0", "0.0", "0.5", "8.0*angstroms", "9.0*angstroms",
-                "isotropic"
-            ])
-        assert_equal(list(attached_units.keys()), ['sigma', 'epsilon'])
-        assert_equal(list(attached_units.values()), [unit.angstroms, unit.kilocalories_per_mole])
-        # Test attachment of units
-        parameter_attrib = OrderedDict([
-            ('smirks', "[#1:1]"),
-            ('sigma', "1.4870"),
-            ('epsilon', "0.0157"),
-        ])
-        unit_bearing_attrib = _attach_units(parameter_attrib, attached_units)
-        assert_equal(list(unit_bearing_attrib.keys()), ['smirks', 'sigma', 'epsilon'])
-        assert_equal(
-            list(unit_bearing_attrib.values()),
-            ["[#1:1]", 1.4870 * unit.angstroms, 0.0157 * unit.kilocalories_per_mole])
-
-    #
-    # Tests for ForceField construction from XML files (.offxml)
-    #
-
-    def test_create_from_offxml():
-        """Test creation of aa SMIRNOFF ForceField object from offxml files.
-        """
-        # These offxml files are located in package data path, which is automatically installed and searched
-        filenames = [smirnoff99Frosst_offxml_filename, tip3p_offxml_filename]
-        # Create a forcefield from a single offxml
-        forcefield = ForceField(filenames[0])
-        # Create a forcefield from multiple offxml files
-        forcefield = ForceField(filenames)
-        # A generator should work as well
-        forcefield = ForceField(iter(filenames))
-
-    def test_create_from_url():
-        """Test creation of a SMIRNOFF ForceField object from URLs
-        """
-        urls = [
-            'https://raw.githubusercontent.com/openforcefield/openforcefield/master/openforcefield/data/forcefield/smirnoff99Frosst.offxml',
-            'https://raw.githubusercontent.com/openforcefield/openforcefield/master/openforcefield/data/forcefield/tip3p.offxml'
-        ]
-        # Test creation with smirnoff99frosst URL
-        forcefield = ForceField(urls[0])
-        # Test creation with multiple URLs
-        forcefield = ForceField(urls)
-        # A generator should work as well
-        forcefield = ForceField(iter(urls))
-
-    def test_create_from_strings():
-        """Test creation of a SMIRNOFF ForceField object from strings
-        """
-        # Test creation with one string
-        forcefield = ForceField(ffxml_contents)
-        # Test creation with list of strings
-        forcefield = ForceField([ffxml_standard, ffxml_constraints])
-        # Test creation with concatenated XML files
-        forcefield = ForceField(ffxml_standard + ffxml_constraints)
-        # Test creation from other combinations
-        forcefield = ForceField(ffxml_contents_gbsa)
-        forcefield = ForceField(ffxml_contents_chargeincrement)
-        forcefield = ForceField(ffxml_contents_noconstraints)
-
-    def test_create_gbsa():
-        """Test reading of ffxml files with GBSA support.
-        """
-        forcefield = ForceField('Frosst_AlkEthOH_GBSA.offxml')
-
-    #
-    # Tests for ForceField writing to XML files
-    #
-
-    def test_save(self):
-        """Test writing and reading of SMIRNOFF in XML format.
-        """
-        forcefield = ForceField(smirnoff99Frosst_offxml_filename)
-        # Write XML to a file
-        with TemporaryDirectory() as tmpdir:
-            offxml_tmpfile = os.path.join(tmpdir, 'forcefield.offxml')
-            forcefield.save(offxml_tmpfile)
-            forcefield2 = ForceField(offxml_tmpfile)
-            assert_forcefields_equal(cls.forcefield, forcefield2,
-                                     "ForceField written to .offxml does not match original ForceField")
-
-    def test_to_xml(self):
-        forcefield = ForceField(smirnoff99Frosst_offxml_filename)
-        # Retrieve XML as a string
-        xml = forcefield.to_xml()
-        # Restore ForceField from XML
-        forcefield2 = ForceField(xml)
-        assert_forcefields_equal(cls.forcefield, forcefield2,
-                                 "ForceField serialized to XML does not match original ForceField")
-
-    def test_deep_copy(self):
-        forcefield = ForceField(smirnoff99Frosst_offxml_filename)
-        # Deep copy
-        forcefield2 = copy.deepcopy(cls.forcefield)
-        assert_forcefields_equal(cls.forcefield, forcefield2,
-                                 "ForceField deep copy does not match original ForceField")
-
-    def test_serialize(self):
-        forcefield = ForceField(smirnoff99Frosst_offxml_filename)
-        # Serialize/deserialize
-        serialized_forcefield = cls.forcefield.__getstate__()
-        forcefield2 = ForceField.__setstate__(serialized_forcefield)
-        assert_forcefields_equal(cls.forcefield, forcefield2,
-                                 "Deserialized serialized ForceField does not match original ForceField")
-
 
 class TestApplyForceField:
     """Test the use of ForceField to parameterize Topology objects.
@@ -552,6 +263,7 @@ class TestApplyForceField:
         topology = Topology.from_molecules(molecule)
         system = forcefield.create_system(topology)
         check_energy_is_finite(system, molecule.positions)
+        # TODO: Add check_energy_is_finite to test_FF.test_paramerize_ethanol
 
     def check_system_creation_from_topology(forcefield, topology, positions):
         """
@@ -588,110 +300,6 @@ class TestApplyForceField:
             f.description = 'Testing {} parameter assignment using molecule {}'.format(offxml_filename, molecule.name)
             yield f
 
-    def test_AlkEthOH_smirnoff99Frosst(self):
-        """Test parameter assignment using smirnoff99Frosst on AlkEthOH.
-        """
-        check_parameter_assignment(
-            offxml_filename=smirnoff99Frosst_offxml_filename, molecules_filename=AlkEthOH_molecules_filename)
-
-    def test_MiniDrugBank_smirnoff99Frosst(self):
-        """Test parameter assignment using smirnoff99Frosst on MiniDrugBank.
-        """
-        check_parameter_assignment(
-            offxml_filename=smirnoff99Frosst_offxml_filename, molecules_filename=MiniDrugBank_molecules_filename)
-
-    def test_electrostatics_options(self):
-        """Test parameter assignment using smirnoff99Frosst on laromustine with various long-range electrostatics options.
-        """
-        molecules_filename = get_data_filename('molecules/laromustine_tripos.mol2')
-        molecule = openforcefield.topology.Molecule.from_file(molecules_filename)
-        forcefield = ForceField([smirnoff99Frosst_offxml_filename, chargeincrement_offxml_filename])
-        for method in ['PME', 'reaction-field', 'Coulomb']:
-            # Change electrostatics method
-            forcefield.forces['Electrostatics'].method = method
-            f = partial(check_system_creation_from_molecule, forcefield, molecule)
-            f.description = 'Testing {} parameter assignment using molecule {}'.format(offxml_filename, molecule.name)
-            yield f
-
-    def test_chargeincrement(self):
-        """Test parameter assignment using smirnoff99Frosst on laromustine with ChargeIncrementModel.
-        """
-        molecules_filename = get_data_filename('molecules/laromustine_tripos.mol2')
-        molecule = openforcefield.topology.Molecule.from_file(molecules_filename)
-        forcefield = ForceField(['smirnoff99Frosst.offxml', 'chargeincrement-test'])
-        check_system_creation_from_molecule(forcefield, molecule)
-
-    def test_create_system_molecules_parmatfrosst_gbsa(self):
-        """Test creation of a System object from small molecules to test parm@frosst forcefield with GBSA support.
-        """
-        molecules_filename = get_data_filename('molecules/AlkEthOH_test_filt1_tripos.mol2')
-        check_parameter_assignment(
-            offxml_filename='Frosst_AlkEthOH_GBSA.offxml', molecules_filename=molecules_filename)
-
-    def test_mixed_solvent_boxes(self):
-        """Test creation of System from boxes of mixed solvents.
-        """
-        forcefield = ForceField('smirnoff99frosst.offxml')
-        # Read unique molecules represented in these systems
-        monomers = ['water', 'cyclohexane', 'ethanol', 'propane', 'methane', 'butanol']
-        filenames = [get_data_filename(os.path.join('systems', 'monomers', monomer + '.sdf')) for monomer in monomers]
-        unique_molecules = [openforcefield.topology.Molecule.from_file(filename) for filename in filenames]
-        if verbose: print('%d reference molecules loaded' % len(molecules))
-
-        # Create Topology objects for systems
-        boxes = [
-            'ethanol_water.pdb', 'cyclohexane_water.pdb', 'cyclohexane_ethanol_0.4_0.6.pdb',
-            'propane_methane_butanol_0.2_0.3_0.5.pdb'
-        ]
-        from simtk.openmm.app import PDBFile
-        for box in boxes:
-            filename = get_data_filename(os.path.join('systems', 'packmol_boxes', box))
-            pdbfile = PDBFile(filename)
-            topology = Topology.from_openmm_topology(pdbfile.topology, unique_molecules)
-            f = partial(check_system_creation_from_topology, forcefield, topology, pdbfile.positions)
-            f.description = 'Test application of smirnoff99Frosst parameters to %s' % (box)
-            yield f
-
-
-class TestForceFieldEnergies:
-    """Compare SMIRNOFF energies with AMBER.
-    """
-
-    def test_smirnoff_energies_vs_parmatfrosst(self):
-        """Test evaluation of energies from parm@frosst ffxml files versus energies of equivalent systems.
-        """
-        prefix = 'AlkEthOH_'
-        molecule_names = ['r118', 'r12', 'c1161', 'r0', 'c100', 'c38', 'c1266']
-
-        # Load special parm@frosst with parm99/parm@frosst bugs re-added for testing
-        forcefield = ForceField('Frosst_AlkEthOH_parmAtFrosst.offxml')
-
-        # Loop over molecules, load OEMols and prep for comparison/do comparison
-        for molecule_name in molecule_names:
-            filename_prefix = os.path.join('molecules', prefix + molecule_name)
-
-            # Create OpenMM System from AMBER prmtop/inpcrd
-            prmtop = app.AmberPrmtopFile(get_data_filename(filename_prefix + '.top'))
-            inpcrd = app.AmberInpcrdFile(get_data_filename(filename_prefix + '.crd'))
-            openmm_topology = prmtop.topology
-            amber_system = prmtop.createSystem(nonbondedMethod=app.NoCutoff, constraints=None, implicitSolvent=None)
-            positions = inpcrd.getPositions()
-
-            # Create openforcefield Topology
-            unique_molecules = Molecule.from_file(get_data_filename(filename_prefix + '.mol2'))
-            openforcefield_topology = Topology.from_openmm(prmtop.topology, unique_molecules)
-            smirnoff_system = forcefield.create_system(openforcefield_topology)
-
-            # Compare energies
-            compare_system_energies(
-                openmm_topology,
-                openmm_topology,
-                amber_system,
-                smirnoff_system,
-                positions,
-                verbose=True,
-                skip_assert=False,
-                skip_improper=False)
 
     def test_partial_bondorder(verbose=False):
         """Test energies of a molecule which activates partial bond order code."""
